todolist edit, delete & update added

+++ todolist/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import redirect, render
 from django.contrib import messages
+import todolist
 from todolist.forms import todoform
 from todolist.models import todolists
 
@@ -22,16 +23,32 @@ def addtodolist(request):
     messages.warning(request, "Something went wrong please try again")
     return redirect('userhome_todo')
 
-def status(request, id):
+def edit_todo(request, id):
+  todolist = todolists.objects.get(id=id)
+  if request.method == 'POST':
+    newtodo = request.POST['newtodo']
+    todolist.todo = newtodo
+    todolist.save()
+    messages.warning(request, "Todolist updated")
+    return redirect('userhome_todo')
+  else:
+    return render(request, "edit_todolist.html",{'todolist':todolist}) 
+
+
+def status_update(request, id):
   todolist = todolists.objects.get(id=id)
   current_status = todolist.todo_completed
-  todolist.todo_completed = not current_status
+  todolist.todo_completed = not current_status 
   todolist.save()
-  return redirect('userhome_todo')
+  messages.warning(request, "Status updated")
+  return redirect('userhome_todo') 
+
 
 def delete_todo(request, id):
   todolist = todolists.objects.get(id=id)
-  todolist.delete()
-  messages.warning(request, "Todolist deleted")
-  return redirect('userhome_todo')
-  
+  if request.method == 'POST':
+    todolist.delete()
+    messages.warning(request, "Todolist deleted")
+    return redirect('userhome_todo')
+  else:
+    return render(request, "confirm_delete.html",{'todolist':todolist})
